feat(extension,setup): renaming ledger components and finalizing import flow
