--- apps/extension/src/Setup/Ledger/Steps/Confirmation/Confirmation.tsx
@@ -1,3 +1,0 @@
-export const Confirmation = (): JSX.Element => {
-  return <></>;
-};
