+++ apps/extension/src/Setup/Ledger/Steps/LedgerConfirmation/index.ts
@@ -1,0 +1,1 @@
+export * from "./LedgerConfirmation";
